+++ test-d/Elements/Element-Types.test-d.ts
@@ -25,7 +25,7 @@ import { expectAssignable, expectType } from 'tsd';
 import lng from '../../index.js';
 import { CompileComponentTemplateSpecType } from '../../src/application/Component.mjs';
 import { SignalMapType } from '../../src/internalTypes.mjs';
-import { InlineElement, IsLooseTemplateSpec, SmoothTemplate, TaggedElements, TemplateSpecRefs, TransformPossibleElement, TransitionsTemplate } from '../../src/tree/Element.mjs';
+import { InlineElement, IsLooseTemplateSpec, SmoothTemplate, TemplateSpecTags, TemplateSpecRefs, TransformPossibleElement, TransitionsTemplate } from '../../src/tree/Element.mjs';
 
 export interface TestTemplateSpec extends lng.Component.TemplateSpec {
   prop1: string;
@@ -302,11 +302,11 @@ function TemplateSpecRefs_Test() {
 }
 
 //
-// TaggedElements
+// TemplateSpecTags
 //
-function TaggedElements_Test() {
+function TemplateSpecTags_Test() {
   /// Strong template specs returns flat tag path map
-  type T1000 = TaggedElements<TestTemplateSpec>;
+  type T1000 = TemplateSpecTags<TestTemplateSpec>;
   expectType<{
     'MyStrongElement_InlineEmpty': lng.Element<InlineElement<{}>>;
     'MyStrongElement_InlineEmpty2': lng.Element<InlineElement<object>>;
@@ -323,8 +323,23 @@ function TaggedElements_Test() {
   }>({} as T1000);
 
   /// Loose template specs return empty object type
-  type T2000 = TaggedElements<TestTemplateSpec & lng.Component.TemplateSpecLoose>;
-  expectType<{}>({} as T2000);
+  type T2000 = TemplateSpecTags<TestTemplateSpec & lng.Component.TemplateSpecLoose>;
+  expectType<{
+    'MyStrongElement_InlineEmpty': lng.Element<InlineElement<{}>>;
+    'MyStrongElement_InlineEmpty2': lng.Element<InlineElement<object>>;
+    'MyStrongElement_ExplicitType': lng.Element<lng.Element.TemplateSpec>;
+    'MyLooseElement': lng.Element;
+    'MyListComponent': lng.components.ListComponent;
+    'MyStrongElement_InlineChildren': lng.Element<InlineElement<TestTemplateSpec['MyStrongElement_InlineChildren']>>;
+    'MyStrongElement_InlineChildren.Child1': lng.Element<InlineElement<{}>>;
+    'MyStrongElement_InlineChildren.Child2': lng.Element<InlineElement<object>>;
+    'MyStrongElement_InlineChildren.Child3': lng.Element<lng.Element.TemplateSpec>;
+    'MyStrongElement_InlineChildren.Child4': lng.components.ListComponent;
+    'MyStrongElement_InlineChildren.Child5': lng.Element<InlineElement<TestTemplateSpec['MyStrongElement_InlineChildren']['Child5']>>;
+    'MyStrongElement_InlineChildren.Child5.GrandChild1': lng.Element<InlineElement<object>>;
+  } & {
+    [s: string]: any;
+  }>({} as T2000);
 }
 
 //